requirements.txt
### Requirements without specified version ###
discord.py
python-dotenv
numpy
